third_party/blink/web_tests/fast/forms/suggested-value-password-reveal-expected.txt

This test is setting suggested values on password fields,
and whether they should be revealed.
| <input>
|   id="password1"
|   type="password"
|   value="initial value"
|   this.value="initial value"
|   <shadow:root>
|     <div>
|       id="placeholder"
|       pseudo="-internal-input-suggested"
|       style="display: block !important; user-select: none !important;"
|       shadow:pseudoId="-internal-input-suggested"
|       "suggested value"
|     <div>
|       "initial value"
| <input>
|   id="password2"
|   type="password"
|   value="initial value"
|   this.value="initial value"
|   <shadow:root>
|     <div>
|       class="reveal-password"
|       id="placeholder"
|       pseudo="-internal-input-suggested"
|       style="display: block !important; user-select: none !important;"
|       shadow:pseudoId="-internal-input-suggested"
|       "suggested value"
|     <div>
|       "initial value"
| "password_with_masked_suggestion.value: initial value"
| "internals.suggestedValue(password_with_masked_suggestion): suggested value"
| "password_with_revealed_suggestion.value: initial value"
| "internals.suggestedValue(password_with_revealed_suggestion): suggested value"
